feat: 🎸 use closure to contain game, add movement and sliding
This encapsulates all of the game logic into an anonymous function. The
function is called and public properties and functions are exported to a
variable G. Exported functions can then be called with G.func() (e.g.
`G.movePlayer(1, 0)`).
This change also adds the moving and sliding logic for the player.